evaluation/requirements.txt
# Core dependencies for Mem0 LOCOMO benchmarking
mem0ai>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Processing and utilities
jinja2>=3.1.0
tqdm>=4.65.0
nltk>=3.8.0

# Metrics calculation
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
bert-score>=0.3.13
rouge-score>=0.1.2
sentence-transformers>=2.2.0

# LangChain and memory frameworks
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20
langmem>=0.0.1

# Graph databases
neo4j>=5.0.0

# Other dudes 
zep-python>=2.0.0

